+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 
 
+## [1.6.1] - 2021-07-01
+
+### Changed
+
+- Masked the `@` symbol in additional information text in the Discord message to the
+  SRP team to not accidentally ping in case a Discord ping was copied in there
+
+
 ## [1.6.0] - 2021-06-08
 
 ### Changed